Module-11-LangChain-Framework/09_Multiuser_Chatbot/requirements.txt
gradio
langchain-openai
python-dotenv
langchain-huggingface
langchain-chroma 
langchain-core
